+++ src/main/java/net/thenextlvl/gopaint/brush/standard/BucketBrush.java
@@ -16,31 +16,34 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.thenextlvl.gopaint.brush;
+package net.thenextlvl.gopaint.brush.standard;
 
 import net.thenextlvl.gopaint.api.brush.setting.BrushSettings;
+import net.thenextlvl.gopaint.api.math.ConnectedBlocks;
 import net.thenextlvl.gopaint.api.math.Sphere;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 
+import java.util.List;
 import java.util.stream.Stream;
 
-public class DiscBrush extends CraftBrush {
+public class BucketBrush extends CraftBrush {
 
-    private static final String DESCRIPTION = "Paints blocks in the\n§8same selected axis\n§8from the block you clicked";
-    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvYjFmMjgyNTBkMWU0MjBhNjUxMWIwMzk2NDg2OGZjYTJmNTYzN2UzYWJhNzlmNGExNjNmNGE4ZDYxM2JlIn19fQ==";
-    private static final String NAME = "Disc Brush";
+    private static final String DESCRIPTION = "Paints connected blocks\n§8with the same block type";
+    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNTAxOGI0NTc0OTM5Nzg4YTJhZDU1NTJiOTEyZDY3ODEwNjk4ODhjNTEyMzRhNGExM2VhZGI3ZDRjOTc5YzkzIn19fQ==";
+    private static final String NAME = "Bucket Brush";
 
-    public DiscBrush() {
+    public BucketBrush() {
         super(NAME, DESCRIPTION, HEAD);
     }
 
     @Override
     public void paint(Location location, Player player, BrushSettings brushSettings) {
         performEdit(player, session -> {
-            Stream<Block> blocks = Sphere.getBlocksInRadius(location, brushSettings.getSize(), brushSettings.getAxis(), false);
-            blocks.filter(block -> passesDefaultChecks(brushSettings, player, block))
+            List<Block> blocks = Sphere.getBlocksInRadius(location, brushSettings.getSize(), null, false).toList();
+            Stream<Block> connectedBlocks = ConnectedBlocks.getConnectedBlocks(location, blocks);
+            connectedBlocks.filter(block -> passesDefaultChecks(brushSettings, player, block))
                     .forEach(block -> setBlock(session, block, brushSettings.getRandomBlock()));
         });
     }
