+++ src/main/java/net/thenextlvl/gopaint/brush/standard/FractureBrush.java
@@ -16,34 +16,42 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.thenextlvl.gopaint.brush;
+package net.thenextlvl.gopaint.brush.standard;
 
 import net.thenextlvl.gopaint.api.brush.setting.BrushSettings;
-import net.thenextlvl.gopaint.api.math.ConnectedBlocks;
+import net.thenextlvl.gopaint.api.math.Height;
 import net.thenextlvl.gopaint.api.math.Sphere;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 
-import java.util.List;
 import java.util.stream.Stream;
 
-public class BucketBrush extends CraftBrush {
+public class FractureBrush extends CraftBrush {
 
-    private static final String DESCRIPTION = "Paints connected blocks\n§8with the same block type";
-    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNTAxOGI0NTc0OTM5Nzg4YTJhZDU1NTJiOTEyZDY3ODEwNjk4ODhjNTEyMzRhNGExM2VhZGI3ZDRjOTc5YzkzIn19fQ==";
-    private static final String NAME = "Bucket Brush";
+    private static final String DESCRIPTION = "Places blocks in cracks/fisures";
+    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvMjNkZjczZWVlNjIyNGM1YzVkOTQ4ZDJhMzQ1ZGUyNWYyMDhjYmQ5YWY3MTA4Y2UxZTFiNjFhNTg2ZGU5OGIyIn19fQ==";
+    private static final String NAME = "Fracture Brush";
 
-    public BucketBrush() {
+    public FractureBrush() {
         super(NAME, DESCRIPTION, HEAD);
     }
 
     @Override
     public void paint(Location location, Player player, BrushSettings brushSettings) {
         performEdit(player, session -> {
-            List<Block> blocks = Sphere.getBlocksInRadius(location, brushSettings.getSize(), null, false).toList();
-            Stream<Block> connectedBlocks = ConnectedBlocks.getConnectedBlocks(location, blocks);
-            connectedBlocks.filter(block -> passesDefaultChecks(brushSettings, player, block))
+            Stream<Block> blocks = Sphere.getBlocksInRadius(location, brushSettings.getSize(), null, false);
+            blocks.filter(block -> passesMaskCheck(brushSettings, block))
+                    .filter(block -> Height.getAverageHeightDiffFracture(
+                            block.getLocation(),
+                            Height.getNearestNonEmptyBlock(block.getLocation()),
+                            1
+                    ) >= 0.1)
+                    .filter(block -> Height.getAverageHeightDiffFracture(
+                            block.getLocation(),
+                            Height.getNearestNonEmptyBlock(block.getLocation()),
+                            brushSettings.getFractureDistance()
+                    ) >= 0.1)
                     .forEach(block -> setBlock(session, block, brushSettings.getRandomBlock()));
         });
     }
