+++ src/main/java/net/thenextlvl/gopaint/brush/standard/SphereBrush.java
@@ -16,10 +16,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package net.thenextlvl.gopaint.brush;
+package net.thenextlvl.gopaint.brush.standard;
 
 import net.thenextlvl.gopaint.api.brush.setting.BrushSettings;
-import net.thenextlvl.gopaint.api.math.Height;
 import net.thenextlvl.gopaint.api.math.Sphere;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
@@ -27,13 +26,13 @@ import org.bukkit.entity.Player;
 
 import java.util.stream.Stream;
 
-public class FractureBrush extends CraftBrush {
+public class SphereBrush extends CraftBrush {
 
-    private static final String DESCRIPTION = "Places blocks in cracks/fisures";
-    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvMjNkZjczZWVlNjIyNGM1YzVkOTQ4ZDJhMzQ1ZGUyNWYyMDhjYmQ5YWY3MTA4Y2UxZTFiNjFhNTg2ZGU5OGIyIn19fQ==";
-    private static final String NAME = "Fracture Brush";
+    private static final String DESCRIPTION = "Regular sphere brush";
+    private static final String HEAD = "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvZmU5OGY0ODU2MDE0N2MwYTJkNGVkYzE3ZjZkOTg1ZThlYjVkOTRiZDcyZmM2MDc0NGE1YThmMmQ5MDVhMTgifX19";
+    private static final String NAME = "Sphere Brush";
 
-    public FractureBrush() {
+    public SphereBrush() {
         super(NAME, DESCRIPTION, HEAD);
     }
 
@@ -41,17 +40,7 @@ public class FractureBrush extends CraftBrush {
     public void paint(Location location, Player player, BrushSettings brushSettings) {
         performEdit(player, session -> {
             Stream<Block> blocks = Sphere.getBlocksInRadius(location, brushSettings.getSize(), null, false);
-            blocks.filter(block -> passesMaskCheck(brushSettings, block))
-                    .filter(block -> Height.getAverageHeightDiffFracture(
-                            block.getLocation(),
-                            Height.getNearestNonEmptyBlock(block.getLocation()),
-                            1
-                    ) >= 0.1)
-                    .filter(block -> Height.getAverageHeightDiffFracture(
-                            block.getLocation(),
-                            Height.getNearestNonEmptyBlock(block.getLocation()),
-                            brushSettings.getFractureDistance()
-                    ) >= 0.1)
+            blocks.filter(block -> passesDefaultChecks(brushSettings, player, block))
                     .forEach(block -> setBlock(session, block, brushSettings.getRandomBlock()));
         });
     }
